exact_requirements.txt
absl-py==1.2.0
alabaster==0.7.12
#-e git+https://github.com/matt-peters/allennlp.git@2d7ba1cb108428aaffe2dce875648253b44cb5ba#egg=allennlp
asttokens==2.1.0
attrs==22.1.0
awscli==1.24.10
Babel==2.11.0
backcall==0.2.0
blis==0.7.8
boto3==1.23.10
botocore==1.26.10
catalogue==2.0.8
#certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
confection==0.0.1
conllu==0.11
contextlib2==21.6.0
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.6
cytoolz==0.9.0.1
decorator==5.1.1
dill==0.2.9
docutils==0.16
dynaconf==3.1.9
editdistance==0.6.1
einops==0.4.1
executing==1.2.0
filelock==3.8.0
flaky==3.7.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.37.1
ftfy==6.1.1
gevent==22.10.2
greenlet==2.0.0
h5py==3.4.0
huggingface-hub==0.9.1
idna==3.4
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonnet==0.19.1
#-e git+ssh://git@gitee.com/le0jc/vit-bert-bottle-classification.git@233c9196d7959fa2e54a5b520f7a41761912ee35#egg=kb&subdirectory=kb
kiwisolver==1.4.4
langcodes==3.2.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
#mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
ml-collections==0.1.1
moto==4.0.9
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.7
nltk==3.6.3
#numpy @ file:///croot/numpy_and_numpy_base_1665773175172/work
numpydoc==1.4.0
overrides==3.1.0
opencv-python==4.7.0.68
packaging==21.0
parsimonious==0.8.1
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plac==0.9.6
pluggy==1.0.0
preshed==2.0.1
prompt-toolkit==3.0.30
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2022.6
PyYAML==5.4.1
regex==2018.1.10
requests==2.27.1
responses==0.20.0
rsa==4.7.2
s3transfer==0.5.2
scikit-learn==0.22.1
scipy==1.9.1
#six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
snowballstemmer==2.1.0
spacy==2.0.18
spacy-legacy==3.0.9
spacy-loggers==1.0.3
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
srsly==2.4.4
stack-data==0.6.0
tensorboardX==2.5.1
thinc==6.12.1
timm==0.6.7
tokenizers==0.12.1
tomli==2.0.1
toolz==0.12.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.63.1
traitlets==5.5.0
transformers==4.22.2
typer==0.4.2
typing-utils==0.1.0
#typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
ujson==5.3.0
Unidecode==1.3.4
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
word2number==1.1
wrapt==1.10.11
xmltodict==0.13.0
yapf==0.32.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
